Refactor Coalesce and Normalizable.
