+++ connector/src/main/scala/quasar/qscript/TTypes.scala
@@ -16,15 +16,9 @@
 
 package quasar.qscript
 
-import scala.Predef.$conforms
-import quasar.Predef._
-import quasar.contrib.matryoshka._
-import quasar.ejson.EJson
 import quasar.fp._
-import quasar.fp.ski._
 
-import matryoshka._
-import scalaz._, Scalaz._
+import scalaz._
 
 /** Centralizes the knowledge of T[_[_]] as well as certain
  *  type classes required for many operations. This is for
@@ -48,7 +42,6 @@ trait TTypes[T[_[_]]] {
 }
 
 object TTypes {
-  def normalizable[T[_[_]] : Recursive : Corecursive : EqualT : ShowT] = new NormalizableT[T]
   def simplifiableProjection[T[_[_]]]                                  = new SimplifiableProjectionT[T]
 }
 
@@ -100,137 +93,5 @@ class SimplifiableProjectionT[T[_[_]]] extends TTypes[T] {
         ej.combine
       )
     )
-  )
-}
-
-// ShowT is needed for debugging
-class NormalizableT[T[_[_]] : Recursive : Corecursive : EqualT : ShowT] extends TTypes[T] {
-  import Normalizable._
-  lazy val rewrite = new Rewrite[T]
-
-  def freeTC(free: FreeQS): FreeQS = {
-    freeTransCata[T, QScriptTotal, QScriptTotal, Hole, Hole](free)(
-      liftCo(rewrite.normalizeCoEnv[QScriptTotal])
-    )
-  }
-
-  def freeTCEq(free: FreeQS): Option[FreeQS] = {
-    val freeNormalized = freeTC(free)
-    (free ≠ freeNormalized).option(freeNormalized)
-  }
-
-  def freeMFEq[A: Equal](fm: Free[MapFunc, A]): Option[Free[MapFunc, A]] = {
-    val fmNormalized = freeMF[A](fm)
-    (fm ≠ fmNormalized).option(fmNormalized)
-  }
-
-  def freeMF[A](fm: Free[MapFunc, A]): Free[MapFunc, A] =
-    freeTransCata[T, MapFunc, MapFunc, A, A](fm)(MapFunc.normalize[T, A])
-
-  def makeNorm[A, B, C](
-    lOrig: A, rOrig: B)(
-    left: A => Option[A], right: B => Option[B])(
-    f: (A, B) => C):
-      Option[C] =
-    (left(lOrig), right(rOrig)) match {
-      case (None, None) => None
-      case (l, r)       => f(l.getOrElse(lOrig), r.getOrElse(rOrig)).some
-    }
-
-  def EquiJoin = make(
-    λ[EquiJoin ~> (Option ∘ EquiJoin)#λ](ej =>
-      (freeTCEq(ej.lBranch), freeTCEq(ej.rBranch), freeMFEq(ej.lKey), freeMFEq(ej.rKey), freeMFEq(ej.combine)) match {
-        case (None, None, None, None, None) => None
-        case (lBranchNorm, rBranchNorm, lKeyNorm, rKeyNorm, combineNorm) =>
-          quasar.qscript.EquiJoin(
-            ej.src,
-            lBranchNorm.getOrElse(ej.lBranch),
-            rBranchNorm.getOrElse(ej.rBranch),
-            lKeyNorm.getOrElse(ej.lKey),
-            rKeyNorm.getOrElse(ej.rKey),
-            ej.f,
-            combineNorm.getOrElse(ej.combine)).some
-      }))
-
-  def ThetaJoin = make(
-    λ[ThetaJoin ~> (Option ∘ ThetaJoin)#λ](tj =>
-      (freeTCEq(tj.lBranch), freeTCEq(tj.rBranch), freeMFEq(tj.on), freeMFEq(tj.combine)) match {
-        case (None, None, None, None) => None
-        case (lBranchNorm, rBranchNorm, onNorm, combineNorm) =>
-          quasar.qscript.ThetaJoin(
-            tj.src,
-            lBranchNorm.getOrElse(tj.lBranch),
-            rBranchNorm.getOrElse(tj.rBranch),
-            onNorm.getOrElse(tj.on),
-            tj.f,
-            combineNorm.getOrElse(tj.combine)).some
-      }))
-
-  def QScriptCore = {
-    // NB: all single-bucket reductions should reduce on `null`
-    def normalizeBucket(bucket: FreeMap): FreeMap = bucket.resume.fold({
-      case MapFuncs.Constant(_) => MapFuncs.NullLit[T, Hole]()
-      case _                    => bucket
-    }, κ(bucket))
-
-    make(λ[QScriptCore ~> (Option ∘ QScriptCore)#λ] {
-      case Reduce(src, bucket, reducers, repair) => {
-        val reducersOpt: List[Option[ReduceFunc[FreeMap]]] =
-          reducers.map(_.traverse(freeMFEq[Hole](_)))
-
-        val reducersNormOpt: Option[List[ReduceFunc[FreeMap]]] =
-          (!reducersOpt.map(_.toList).flatten.isEmpty).option(
-            Zip[List].zipWith(reducersOpt, reducers)(_.getOrElse(_)))
-
-        val bucketNormOpt: Option[FreeMap] = freeMFEq(bucket)
-
-        val bucketNormConst: Option[FreeMap] =
-          bucketNormOpt.getOrElse(bucket).resume.fold({
-            case MapFuncs.Constant(ej) =>
-              (!EJson.isNull(ej)).option(MapFuncs.NullLit[T, Hole]())
-            case _ => bucketNormOpt
-          }, κ(bucketNormOpt))
-
-        (bucketNormConst, reducersNormOpt, freeMFEq(repair)) match {
-          case (None, None, None) =>
-            None
-          case (bucketNorm, reducersNorm, repairNorm)  =>
-            Reduce(
-              src,
-              bucketNorm.getOrElse(bucket),
-              reducersNorm.getOrElse(reducers),
-              repairNorm.getOrElse(repair)).some
-        }
-      }
-
-      case Sort(src, bucket, order) => {
-        val orderOpt: List[Option[(FreeMap, SortDir)]] =
-          order.map {
-            _.leftMap(freeMFEq(_)) match {
-              case (Some(fm), dir) => Some((fm, dir))
-              case (_, _)          => None
-            }
-          }
-
-        val orderNormOpt: Option[List[(FreeMap, SortDir)]] =
-          (!orderOpt.map(_.toList).flatten.isEmpty).option(
-            Zip[List].zipWith(orderOpt, order)(_.getOrElse(_)))
-
-        makeNorm(bucket, order)(freeMFEq(_), _ => orderNormOpt)(Sort(src, _, _))
-      }
-      case Map(src, f)            => freeMFEq(f).map(Map(src, _))
-      case LeftShift(src, s, r)   => makeNorm(s, r)(freeMFEq(_), freeMFEq(_))(LeftShift(src, _, _))
-      case Union(src, l, r)       => makeNorm(l, r)(freeTCEq(_), freeTCEq(_))(Union(src, _, _))
-      case Filter(src, f)         => freeMFEq(f).map(Filter(src, _))
-      case Subset(src, from, sel, count) => makeNorm(from, count)(freeTCEq(_), freeTCEq(_))(Subset(src, _, sel, _))
-      case Unreferenced()         => None
-    })
-  }
-
-  def ProjectBucket = make(
-    λ[ProjectBucket ~> (Option ∘ ProjectBucket)#λ] {
-      case BucketField(a, v, f) => makeNorm(v, f)(freeMFEq(_), freeMFEq(_))(BucketField(a, _, _))
-      case BucketIndex(a, v, i) => makeNorm(v, i)(freeMFEq(_), freeMFEq(_))(BucketIndex(a, _, _))
-    }
   )
 }
